refactoring of the methods
